Implemented executing the reader where zero rows are possible
Before, an exception would be thrown if zero rows were returned. For functions that list multiple items, zero results are a possibility. This should not throw an exception as it is not invalid behaviour.

The `ExecuteReaderCanHaveZeroRowsAsync()` function could be replaced by simply calling `command.ExecuteReaderAsync()`. At the moment this function is used because we might want to add checks later.